Contiguousmem (#6)
* started kmeans kernel

* started kmeans kernel

* initialized bgd kmeans kernel - added cudaMallocs, streams, and currently working on rand portion and onward

* added cudaMemCpy, curand and random int list

* INCOMPLETE contiguous rgb block structs for images

* updates

* updates

* updates

* almost done

* compiling!

* kmeans compile and runs

* fixed edit for rand() idx selection. Still wrong results

* kmeans hardcoded centroid array

* whyyy

* shared memory

* altered kmeans id and centroids

* cleaned up code

* cleaned up code

* added timing for gpu code - threadperblock exp

* fixed itming

* fixed itming

* fixed itming

* fixed itming

* results for shared memory for image exp

---------

+++ benchmark/grabcut.h
@@ -60,7 +60,7 @@ typedef double* weight_t;
 
 // For kmeans
 typedef struct {
-    float r, g, b;
+    uint8_t r, g, b;
 } Centroid;
 
 float distance_squared(pixel_t p, Centroid c) {
